Relocate EKS provider code
