+++ internal/cluster/distribution/eks/eksprovider/workflow/create_iam_roles_activity.go
