+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_cluster_from_store_activity.go
