+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_cluster_user_access_key_activity.go
