+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_control_plane_activity.go
