+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_infra.go
