+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_ssh_key_activity.go
