+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_stack_activity.go
