+++ internal/cluster/distribution/eks/eksprovider/workflow/get_nodepool_stacks.go
