+++ internal/cluster/distribution/eks/eksprovider/workflow/get_orphan_nics_activity.go
