+++ internal/cluster/distribution/eks/eksprovider/workflow/get_subnets_details_activity.go
