+++ internal/cluster/distribution/eks/eksprovider/workflow/get_vpc_config_activity.go
