+++ internal/cluster/distribution/eks/eksprovider/workflow/save_network_details_activity.go
